Refactoring the package name (#38)
* Refactoring the package name
