+++ src/main/java/org/ballerinalang/integration/tests/core/commons/DeploymentConfigurationReader.java
@@ -16,12 +16,12 @@
 *  under the License.
 */
 
-package org.ballerina.integration.tests.core.commons;
+package org.ballerinalang.integration.tests.core.commons;
 
 import org.apache.commons.logging.Log;
 import org.apache.commons.logging.LogFactory;
-import org.ballerina.integration.tests.core.FrameworkConstants;
-import org.ballerina.integration.tests.core.beans.Deployment;
+import org.ballerinalang.integration.tests.core.FrameworkConstants;
+import org.ballerinalang.integration.tests.core.beans.Deployment;
 import org.yaml.snakeyaml.Yaml;
 
 import java.io.FileInputStream;
