+++ src/main/java/org/ballerinalang/integration/tests/core/commons/GenericYamlParser.java
@@ -16,7 +16,7 @@
 *  under the License.
 */
 
-package org.ballerina.integration.tests.core.commons;
+package org.ballerinalang.integration.tests.core.commons;
 
 import org.apache.commons.logging.Log;
 import org.apache.commons.logging.LogFactory;
