+++ src/test/resources/docker/create-docker.sh
@@ -13,7 +13,7 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-# This file is triggered from init.sh with the Docker Entry point
+# This file is triggered from Jenkins pre-build Step
 
 current_location=${WORKSPACE}
 echo "Current location is set to : " ${current_location}
